pyproject.toml
[tool.poetry]
name = "kukasaataa"
version = "0.1.0"
description = ""
authors = ["Taina Hilliaho <taina.hilliaho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-cors = "^5.0.0"
requests = "^2.32.3"
invoke = "^2.2.0"
pymupdf = "^1.24.14"
fitz = "^0.0.1.dev2"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
